pyproject.toml
[project]
name = "resto-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "starlette>=0.47.3",
    "uvicorn>=0.35.0",
]
